+++ src/pages/programDisaggregations/form/CategoryMapping.tsx
@@ -80,11 +80,11 @@ type CategoryMappingProps = {
     categoryOptionArray: CategoryOption[]
     showSoftDelete: boolean
 }
-export const CategoryMapping = ({
+export const CategoryMapping = React.memo(function CategoryMapping({
     fieldName,
     categoryOptionArray,
     showSoftDelete = true,
-}: CategoryMappingProps) => {
+}: CategoryMappingProps) {
     const categoryMapping = useField(fieldName)
     const categoryMappingOnChange = categoryMapping?.input?.onChange
     const categoryMappingValue = categoryMapping?.input?.value
@@ -184,7 +184,7 @@ export const CategoryMapping = ({
             ))}
         </CategoryMappingWrapper>
     )
-}
+})
 
 const CategoryMappingInput = ({
     fieldName,