+++ src/pages/programDisaggregations/form/useUpdateProgramIndicatorMutation.ts
@@ -4,7 +4,7 @@ import { useCallback } from 'react'
 import { parseErrorResponse, SECTIONS_MAP, usePatchModel } from '../../../lib'
 import { JsonPatchOperation } from '../../../types'
 import {
-    CategoryMappingsRecord,
+    ProgramDisaggregationFormValues,
     ProgramIndicatorMappingsRecord,
 } from './programDisaggregationSchema'
 
@@ -25,15 +25,16 @@ export const useUpdateProgramIndicatorMutation = () => {
                   ])
                 : null
 
-            const newCategoryMappingsIds =
-                programIndicatorMapping?.categoryCombo
-                    ? programIndicatorMapping.categoryCombo.categories.map(
-                          (category) =>
-                              programIndicatorMapping.disaggregation[
-                                  category.id
-                              ]
-                      )
-                    : []
+            const newAttributeCombo = programIndicatorMapping?.attributeCombo
+                ? pick(programIndicatorMapping.attributeCombo, [
+                      'id',
+                      'displayName',
+                  ])
+                : null
+            const newCategoryMappingIds = Object.values(
+                programIndicatorMapping?.disaggregation ?? {}
+            ).concat(Object.values(programIndicatorMapping?.attribute ?? {}))
+
             const newAggregateExportDataElement =
                 programIndicatorMapping?.aggregateExportDataElement ?? null
 
@@ -49,8 +50,13 @@ export const useUpdateProgramIndicatorMutation = () => {
                     },
                     {
                         op: 'replace',
+                        path: '/attributeCombo',
+                        value: newAttributeCombo,
+                    },
+                    {
+                        op: 'replace',
                         path: '/categoryMappingIds',
-                        value: newCategoryMappingsIds,
+                        value: newCategoryMappingIds,
                     },
                     {
                         op: 'replace',
@@ -75,20 +81,11 @@ export const useUpdateProgramMutation = (programId: string) => {
     const patchModel = usePatchModel(programId, SECTIONS_MAP.program.namePlural)
 
     return useCallback(
-        async (
-            categoryMappings: CategoryMappingsRecord & { deleted?: string[] }
-        ) => {
-            const { deleted, ...updatedCategoryMappings } = categoryMappings
-
-            const cleanCatMappings = Object.fromEntries(
-                Object.entries(updatedCategoryMappings).filter(
-                    ([key]) => !deleted?.includes(key)
-                )
-            )
-
-            const newCategoryMappings = Object.values(cleanCatMappings)
+        async ({
+            categoryMappings,
+        }: Pick<ProgramDisaggregationFormValues, 'categoryMappings'>) => {
+            const newCategoryMappings = Object.values(categoryMappings)
                 .flat()
-                .filter((mapping) => !mapping.deleted)
                 .map(({ id, categoryId, mappingName, options }) => ({
                     id,
                     categoryId,